requirements.txt
openai
langchain
pypdf
yt_dlp
pydub
ffmpeg
ffprobe
pydu
python-dotenv
beautifulsoup4
tiktoken
chromadb
lark
